# refactor: improve client executor loop