On top of #4227 

Make client side state machines go brrrr:

* no busy looping/polling anything (good for phone bateries)
* no canceling anything
* parallel state transitions

 No state machine was harmed during the rewrite.