@dpc 2024-02-08
@elsirion I'm afraid until https://github.com/fedimint/fedimint/pull/4250 is figured out, this is making the CI too flaky to land.