@douglaz 2024-02-09
> Make client side state machines go brrrr:
> 
> * no busy looping/polling anything (good for phone bateries)

> * parallel state transitions

The parallel state transitions may lead to duplicated processing. So it may or may not improve phone batteries.

Perhaps we should keep them serial for now? I'm not sure we understand all the possible side effects and new issues that may happen.