@dpc 2024-02-09
> The parallel state transitions may lead to duplicated processing. So it may or may not improve phone batteries.

Not polling on empty executor is the main benefit. The second one is not canceling all pending futures. They key to saving battery is not waking up CPUs unnecessarily, so they can sit in low power mode for as long as possible.

> Perhaps we should keep them serial for now? I'm not sure we understand all the possible side effects and new issues that may happen.


Processing state change in parallel or not is a minor thing, IMO, one or the other. My bet is conflicts here will not be happening a lot, and even if they do, coupe of retries with a backoff and everything should complete anyway in a burst of activity, after which the CPUs can go back to sleep.

I don't want to pesimise right away. If it looks like db conflicts are common, I'll start by introducing a map of `operation_id -> mutex` or `module_id -> mutex` to serialize things touching same parts.

And I think And some conflicts from time to time will be beneficial as they will exercise idempotency of state transition functions.