@douglaz 2024-02-09
@dpc the question is when we are including this in a stable release. If we want to include it in the foreseen future, we should be more conservative, incremental and pessimistic about the changes.