@dpc 2024-02-09
> oh formatting is totally messed up, rustfmt is giving up

The arm of that `match` for `ExecutorLoopEvent::Triggered` is the only part affected, it seems. Beats me why, and it's hard to tweak it too much without breaking it.

I don't think I care too much, unless you see some particular eyesores.