@dpc 2024-02-09
BTW. We do have a big red `warn!` on autocommit fail, so we'll see quickly if parallel state machine transfer functions are conflicting.