@maan2003 2024-02-09
I created #4291 to make sure this doesn't break wasm