@bradleystachurski 2024-02-14
Looks like a few of the rust unit tests are timing out (flake?).

```
TIMEOUT [ 480.039s] fedimint-ln-gateway::gatewayd-integration-tests
TIMEOUT [ 480.092s] fedimint-wallet-tests::fedimint_wallet_tests peg_out_fail_refund
```

I think we can disable `rust_unit_tests` in backwards-compatibility tests since it's not using binaries from previous versions.