@dpc 2024-02-15
Seems like this change improves latency tests 2 times.

```
before 1st try:
================= RESULTS ==================
REISSUE: min: 1.5s, avg: 1.6s, median: 1.6s, p90: 1.6s, max: 1.6s, sum: 47.0s
LN SEND: min: 2.9s, avg: 3.0s, median: 2.9s, p90: 3.0s, max: 3.8s, sum: 59.0s
LN RECV: min: 1.4s, avg: 1.8s, median: 1.5s, p90: 2.4s, max: 2.4s, sum: 35.5s
FM PAY: min: 3.5s, avg: 3.6s, median: 3.6s, p90: 3.6s, max: 3.6s, sum: 71.3s


before, 2nd try:
================= RESULTS ==================
REISSUE: min: 1.5s, avg: 1.6s, median: 1.6s, p90: 1.6s, max: 1.6s, sum: 47.0s
LN SEND: min: 2.8s, avg: 2.9s, median: 2.9s, p90: 2.9s, max: 3.7s, sum: 58.7s
LN RECV: min: 1.4s, avg: 1.8s, median: 1.5s, p90: 2.4s, max: 2.4s, sum: 35.0s
FM PAY: min: 3.5s, avg: 3.8s, median: 3.6s, p90: 3.6s, max: 8.6s, sum: 76.3s


after, 1st try:
================= RESULTS ==================
REISSUE: min: 0.5s, avg: 0.6s, median: 0.6s, p90: 0.6s, max: 0.6s, sum: 16.6s
LN SEND: min: 1.3s, avg: 1.5s, median: 1.5s, p90: 1.7s, max: 2.0s, sum: 30.0s
LN RECV: min: 1.3s, avg: 1.6s, median: 1.6s, p90: 1.8s, max: 1.9s, sum: 32.4s
FM PAY: min: 1.5s, avg: 1.6s, median: 1.6s, p90: 1.6s, max: 1.8s, sum: 32.1s


after, 2nd try:
================= RESULTS ==================
REISSUE: min: 0.5s, avg: 0.6s, median: 0.5s, p90: 0.6s, max: 0.6s, sum: 16.6s
LN SEND: min: 1.3s, avg: 1.5s, median: 1.4s, p90: 1.7s, max: 2.1s, sum: 29.9s
LN RECV: min: 1.3s, avg: 1.6s, median: 1.6s, p90: 1.6s, max: 1.6s, sum: 31.9s
FM PAY: min: 1.5s, avg: 1.6s, median: 1.6s, p90: 1.6s, max: 1.7s, sum: 32.2s
```